+++ CHANGES.md
@@ -5,6 +5,7 @@
 * update description of pycobertura
 * pep8-ify
 * add pep8 tasks for tox and travis
+* diff command returns non-zero exit code if coverage worsened
 
 ## 0.2.1 (2014-12-10)